Name submeshes after MeshInstance instead of Mesh
It seems the phyre Engine compiler mashes all the meshes of each mesh instance together when compiling a .dae file, so the mesh name ends up crazy long.  Since the mesh instance has the name of the first mesh, and only has one mesh, it makes sense that it should be safe to name the submeshes after the mesh instance instead of the mesh.  (To my understanding, a mesh can have more than one instance, but an instance should only have one mesh... I hope.)